Changes in customer adding popup and attachment in all forms

+++ PilotSmithApp.UserInterface/Scripts/custom.js
@@ -682,9 +682,14 @@ function DeleteNow(this_Obj)
 
     }    
 }
-function PaintImages(ID)
+function PaintImages(ID, isApproval)
 {
+    debugger;
     try {
+        if (isApproval == undefined)
+        {
+            isApproval = false;
+        }
         var data = { "ID": ID };
         var ds = {};
         ds = GetDataFromServer("FileUpload/GetAttachments/", data);
@@ -699,7 +704,7 @@ function PaintImages(ID)
                 var filesize = 0;
                 for (var i = 0; i < ds.Records.length; i++) {
                     var html = "";
-                    if (ds.Records[i].IsDocLocked == false)
+                    if (ds.Records[i].IsDocLocked == false )
                         html = '<div class="file-preview-thumbnails">'
                                         + '  <div class="file-preview-frame krajee-default  kv-preview-thumb">'
                                              + ' <div class="kv-file-content">'
@@ -716,24 +721,46 @@ function PaintImages(ID)
                                            + '   </div>'
                                        + '   </div>'
                                       + '</div>'
-                    else
-                    html = '<div class="file-preview-thumbnails">'
-                                        + '  <div class="file-preview-frame krajee-default  kv-preview-thumb">'
-                                             + ' <div class="kv-file-content">'
-                                             + '<a href="/FileUpload/DownloadFile?token=' + ds.Records[i].AttachmentURL + '" style="position: absolute;left: 7%;cursor:pointer;z-index: 900;color: #26a026;"><i class="fa fa-download" aria-hidden="true" ></i></a>'
-                                             + '<a style="position: absolute;right: 0%;cursor:pointer;z-index: 900;color: #dc3939;" ><i class="fa fa-trash-o" aria-hidden="true" onclick="DeleteFile(this);" token="' + ds.Records[i].ID + '"></i></a>'
-                                                 + ' <div class="kv-preview-data file-preview-other-frame">'
-                                                     + ' <div class="file-preview-other">'
-                                                        + '  <span class="file-other-icon">' + validateType(ds.Records[i].FileName) + '</span>'
-                                                     + ' </div>'
-                                                  + '</div>'
-                                              + '</div>'
-                                            + '  <div class="file-thumbnail-footer">'
-                                                 + ' <div class="file-footer-caption" title="' + ds.Records[i].FileName + '">' + ds.Records[i].FileName + '<br> <samp>(' + bytesToSize(ds.Records[i].FileSize) + ')</samp></div>'
+                    else if (ds.Records[i].IsDocLocked == true) {
+                        if (isApproval == true) {
+                            html = '<div class="file-preview-thumbnails">'
+                                                                  + '  <div class="file-preview-frame krajee-default  kv-preview-thumb">'
+                                                                       + ' <div class="kv-file-content">'
+                                                                       + '<a href="/FileUpload/DownloadFile?token=' + ds.Records[i].AttachmentURL + '" style="position: absolute;left: 7%;cursor:pointer;z-index: 900;color: #26a026;"><i class="fa fa-download" aria-hidden="true" ></i></a>'
+                                                                           + ' <div class="kv-preview-data file-preview-other-frame">'
+                                                                               + ' <div class="file-preview-other">'
+                                                                                  + '  <span class="file-other-icon">' + validateType(ds.Records[i].FileName) + '</span>'
+                                                                               + ' </div>'
+                                                                            + '</div>'
+                                                                        + '</div>'
+                                                                      + '  <div class="file-thumbnail-footer">'
+                                                                           + ' <div class="file-footer-caption" title="' + ds.Records[i].FileName + '">' + ds.Records[i].FileName + '<br> <samp>(' + bytesToSize(ds.Records[i].FileSize) + ')</samp></div>'
 
+                                                                     + '   </div>'
+                                                                 + '   </div>'
+                                                                + '</div>'
+                        }
+                        else {
+                            html = '<div class="file-preview-thumbnails">'
+                                            + '  <div class="file-preview-frame krajee-default  kv-preview-thumb">'
+                                                 + ' <div class="kv-file-content">'
+                                                 + '<a href="/FileUpload/DownloadFile?token=' + ds.Records[i].AttachmentURL + '" style="position: absolute;left: 7%;cursor:pointer;z-index: 900;color: #26a026;"><i class="fa fa-download" aria-hidden="true" ></i></a>'
+                                                 + '<a style="position: absolute;right: 0%;cursor:pointer;z-index: 900;color: #dc3939;" ><i class="fa fa-trash-o" aria-hidden="true" onclick="DeleteFile(this);" token="' + ds.Records[i].ID + '"></i></a>'
+                                                     + ' <div class="kv-preview-data file-preview-other-frame">'
+                                                         + ' <div class="file-preview-other">'
+                                                            + '  <span class="file-other-icon">' + validateType(ds.Records[i].FileName) + '</span>'
+                                                         + ' </div>'
+                                                      + '</div>'
+                                                  + '</div>'
+                                                + '  <div class="file-thumbnail-footer">'
+                                                     + ' <div class="file-footer-caption" title="' + ds.Records[i].FileName + '">' + ds.Records[i].FileName + '<br> <samp>(' + bytesToSize(ds.Records[i].FileSize) + ')</samp></div>'
+
+                                               + '   </div>'
                                            + '   </div>'
-                                       + '   </div>'
-                                      + '</div>'
+                                          + '</div>'
+
+                        }
+                    }
                     $('#ExistingPreview').append(html);
                     filesize = filesize + parseInt(ds.Records[i].FileSize);
                 }
